Fixes the UsfmFilterPage routines and other related functions so that the embedded content markers associated with the \x cross reference marker are filtered or unfiltered in unison with \x itself. The embedded content markers associated with the \f footnote and \fe endnote markers are filtered or unfiltered in unison with \f and \fe. In fact, now \f and \fe are always filtered or unfiltered together. If one if filtered the other also is filtered. If one if unfiltered the other is also unfiltered. In support of the above I added a m_crossRefMarkerSet member and a m_footnoteMarkerSet member to the App class, and also added a CleanupFilterMarkerOrphansInString() function that self-heals marker strings loaded from those stored in old xml documents and/or old AI-ProjectConfiguration.aic files. Added more code checking to protect against potential out-of-range array access actions. Added two functions to helpers: AddFilterMarkerToString() and RemoveFilterMarkerFromString() which have the smarts to keep the \x, \f, and \fe filter markers filtered/unfiltered in unison with their associated content markers (and \f in unison with \fe. Added a SetFilterFlagsInIntArray() function to encapsulate the maintenance and consistency of the filter flags array. Improved the CUsfmFilterPageCommon::DoBoxClickedIncludeOrFilterOutDoc() and CUsfmFilterPageCommon::DoBoxClickedIncludeOrFilterOutProj() functions to fix the filtering bugs reported by Bruce on 7Jul12. Corrected some g++ warnings.

+++ source/ProjectPage.cpp
@@ -728,11 +728,11 @@ void CProjectPage::OnWizardPageChanging(wxWizardEvent& event)
 					bool bQueryForSource = FALSE;
 					bool bQueryForTarget = FALSE;
 					bool bQueryForFreeTrans = FALSE;
-					if (errProjects.find(_T("source")) != wxNOT_FOUND)
+					if ((int)errProjects.find(_T("source")) != wxNOT_FOUND)
 						bQueryForSource = TRUE;
-					if (errProjects.find(_T("target")) != wxNOT_FOUND)
+					if ((int)errProjects.find(_T("target")) != wxNOT_FOUND)
 						bQueryForTarget = TRUE;
-					if (errProjects.find(_T("freetrans")) != wxNOT_FOUND)
+					if ((int)errProjects.find(_T("freetrans")) != wxNOT_FOUND)
 						bQueryForFreeTrans = TRUE;
 					wxString caption = _("You may be able to fix this problem if you do the following...");
 					wxString message = _T("");
